[test] Remove some less relevant tests
