+++ integration-tests/src/tests/mod.rs
@@ -2,7 +2,6 @@ mod client;
 mod features;
 mod nearcore;
 mod network;
-mod other_tests;
 mod runtime;
 mod standard_cases;
 mod tools;
