+++ integration-tests/src/tests/tools/mod.rs
@@ -1,4 +1,5 @@
 mod apply_chain_range;
 mod apply_chunk;
+mod dependencies;
 mod state_dump;
 mod state_viewer;
